refactor: Card コンポーネントからカードの枠色を表示する部分を分割し、CardFrame コンポーネントとして新たに定義した
