+++ src/components/Card/index.tsx
@@ -1,6 +1,7 @@
 import { AspectRatio, Box } from '@chakra-ui/react';
 import Image from 'next/image';
 
+import CardFrame from '../CardFrame';
 import type { CardType, ImgSize } from '@/types/cards';
 
 type Props = {
@@ -9,27 +10,8 @@ type Props = {
 };
 
 const Card: React.FC<Props> = ({ card, imgSize }) => {
-  const bgGradients = {
-    R: 'linear(0deg, #8e9ab0, #fefefe)',
-    SR: 'linear(0deg, #e2af38, #fff1cd)',
-    SSR: 'linear(135deg, #fc9edd,#d356f5 30%, #60dbfb 55%,#81f8d3 75%, #dbfe65)',
-  };
-  const outerBorderColors = {
-    R: '#bbb',
-    SR: '#bfa763',
-    SSR: '#95829d',
-  };
-
   return (
-    <Box
-      position='relative'
-      p={{ base: '2px', md: '6px' }}
-      bgGradient={bgGradients[card.card_rarity]}
-      borderRadius={{ base: 8, md: 12 }}
-      borderWidth='1px'
-      borderColor={outerBorderColors[card.card_rarity]}
-      shadow='md'
-    >
+    <CardFrame cardRarity={card.card_rarity}>
       <AspectRatio overflow='hidden' borderRadius={{ base: 6, md: 10 }} ratio={3 / 4}>
         <Image
           layout='fill'
@@ -51,7 +33,7 @@ const Card: React.FC<Props> = ({ card, imgSize }) => {
           alt={`サポートカード "${card.card_name}" のタイプアイコン`}
         />
       </Box>
-    </Box>
+    </CardFrame>
   );
 };
